move them all into utils/replays/hooks instead of components/
