+++ static/app/utils/replays/hooks/useFullscreen.tsx
