+++ static/app/utils/replays/hooks/useRAF.tsx
