+++ static/app/views/replays/detail/focusArea.tsx
@@ -2,11 +2,11 @@ import {useMemo} from 'react';
 
 import EventEntry from 'sentry/components/events/eventEntry';
 import Placeholder from 'sentry/components/placeholder';
-import useActiveTabFromLocation from 'sentry/components/replays/hooks/useActiveTabFromLocation';
 import {useReplayContext} from 'sentry/components/replays/replayContext';
 import TagsTable from 'sentry/components/tagsTable';
 import type {Entry, Event} from 'sentry/types/event';
 import {EntryType} from 'sentry/types/event';
+import useActiveTabFromLocation from 'sentry/utils/replays/hooks/useActiveTabFromLocation';
 import ReplayReader from 'sentry/utils/replays/replayReader';
 import useOrganization from 'sentry/utils/useOrganization';
 import {useRouteContext} from 'sentry/utils/useRouteContext';
